requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-admin-interface==0.17.3
django-ckeditor==6.1.0
django-colorfield==0.4.5
django-cors-headers==3.10.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
djangorestframework==3.12.4
drf-yasg==1.20.0
facebook-sdk==3.1.0
google-api-core==2.1.1
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.2
Pillow==8.4.0
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2021.3
random2==1.0.1
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
django-countries==7.2.1
django-filter==21.1
django-import-export==2.5.0
bs4==0.0.1
django-crontab==0.7.1
numpy~=1.22.3
stripe~=2.68.0

phonenumbers~=8.12.46
beautifulsoup4~=4.10.0
pycountry~=22.3.5